Unify the two featurecalcs into one
